+++ packages/rum/webpack.config.js
@@ -24,93 +24,45 @@
  */
 
 const { join } = require('path')
-const { EnvironmentPlugin } = require('webpack')
-const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer')
-const UglifyJSPlugin = require('uglifyjs-webpack-plugin')
-const { getBabelConfig, BUNDLE_TYPES } = require('../../dev-utils/build')
+const {
+  BUNDLE_TYPES,
+  getWebpackReleaseConfig
+} = require('../../dev-utils/build')
 
 const OUTPUT_DIR = join(__dirname, 'dist', 'bundles')
 const SRC_DIR = join(__dirname, 'src')
-const REPORTS_DIR = join(__dirname, 'reports')
 
 const devConfig = entry => ({
+  ...getWebpackReleaseConfig(BUNDLE_TYPES.BROWSER_DEV),
   entry,
   output: {
     filename: '[name].umd.js',
     path: OUTPUT_DIR,
     library: '[name]',
     libraryTarget: 'umd'
-  },
-  mode: 'development',
-  stats: {
-    assets: true,
-    modules: false
-  },
-  node: false,
-  module: {
-    rules: [
-      {
-        test: /\.js$/,
-        loader: 'babel-loader',
-        options: getBabelConfig(BUNDLE_TYPES.BROWSER_DEV)
-      }
-    ]
-  },
-  plugins: [
-    new EnvironmentPlugin({
-      NODE_ENV: 'development'
-    })
-  ]
+  }
 })
 
 const prodConfig = name => ({
+  ...getWebpackReleaseConfig(BUNDLE_TYPES.BROWSER_PROD, name),
   output: {
     filename: '[name].umd.min.js',
     path: OUTPUT_DIR
-  },
-  mode: 'production',
-  devtool: 'source-map',
-  optimization: {
-    minimizer: [
-      new UglifyJSPlugin({
-        sourceMap: true,
-        extractComments: true
-      })
-    ]
-  },
-  performance: {
-    hints: 'warning',
-    maxAssetSize: 65 * 1024 // 65 kB
-  },
-  plugins: [
-    new EnvironmentPlugin({
-      NODE_ENV: 'production'
-    }),
-    new BundleAnalyzerPlugin({
-      analyzerMode: 'static',
-      reportFilename: join(REPORTS_DIR, `${name}-report.html`),
-      generateStatsFile: true,
-      statsFilename: join(REPORTS_DIR, `${name}-stats.json`),
-      openAnalyzer: false
-    })
-  ]
+  }
 })
 
 const rumDevConfig = devConfig({
   'elastic-apm-rum': join(SRC_DIR, 'index.js')
 })
-
-const rumProdConfig = Object.assign({}, rumDevConfig, prodConfig('apm-rum'))
+const rumProdConfig = { ...rumDevConfig, ...prodConfig('apm-rum') }
 
 const rumOpenTracingDevConfig = devConfig({
   'elastic-apm-opentracing': join(SRC_DIR, 'opentracing.js')
 })
-
-const rumOpenTracingProdConfig = Object.assign(
-  {},
-  rumOpenTracingDevConfig,
-  prodConfig('apm-opentracing')
-)
+const rumOpenTracingProdConfig = {
+  ...rumOpenTracingDevConfig,
+  ...prodConfig('apm-opentracing')
+}
 
 module.exports = [
   rumDevConfig,