+++ packages/rum-react/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/elastic/apm-agent-rum-js/compare/@elastic/apm-rum-react@0.3.2...@elastic/apm-rum-react@0.3.3) (2020-01-15)
+
+**Note:** Version bump only for package @elastic/apm-rum-react
+
+
+
+
+
 ## [0.3.2](https://github.com/elastic/apm-agent-rum-js/compare/@elastic/apm-rum-react@0.3.1...@elastic/apm-rum-react@0.3.2) (2019-11-19)
 
 **Note:** Version bump only for package @elastic/apm-rum-react